requirements.txt
praw~=7.1.0
PyYAML~=5.4.1
pandas~=1.2.1
requests~=2.25.1